// archive/borongan-eService-system-original/multysis-frontend/src/types/socket.types.ts
export interface NewSubscriberPayload {
  id: string;
  firstName: string;
  middleName?: string;
  lastName: string;
  extensionName?: string;
  phoneNumber: string;
  email?: string;
  status?: string;
  type: 'CITIZEN' | 'SUBSCRIBER';
  createdAt: Date | string;
}

export interface SubscriberUpdatePayload {
  subscriberId: string;
  status?: string;
  updatedAt: Date | string;
}